curl get post 请求方式

+++ CURL_Android_Demo/app/src/main/cpp/include/lchttp.h
@@ -8,7 +8,7 @@ struct lchttp_error_t {
 };
 
 struct lchttp_response_t {
-    long code;
+    int code;
     char *header;
     char *text;
 };
